refactor: move comment retrieving and creation functionality from Blog to Comment module
