+++ src/modules/blogs/tests/blogs.controller.spec.ts
@@ -4,7 +4,6 @@ import { BlogsController } from "#/modules/blogs/blogs.controller";
 import { BlogsService } from "../blogs.service";
 import { randomUUID } from "crypto";
 import { Blog } from "../entities/blog.entity";
-import { Comment } from "#/modules/comments/entities/comment.entity";
 
 const app = express();
 
@@ -59,27 +58,6 @@ const mockBlogs = [
 ];
 
 const mockBlog = mockBlogs[0] as Blog;
-
-const mockComments: Comment[] = [
-	{
-		id: "1",
-		content: "Comment 1",
-		blog: { id: "1" },
-		user: { id: "1", username: "davranbek", email: "davranbek@gmail.com" },
-		createdAt: new Date(),
-		updatedAt: new Date(),
-	},
-	{
-		id: "2",
-		content: "Comment 2",
-		blog: { id: "1" },
-		user: { id: "1", username: "davranbek", email: "davranbek@gmail.com" },
-		createdAt: new Date(),
-		updatedAt: new Date(),
-	},
-];
-
-const mockComment = mockComments[0] as Comment;
 
 describe("BlogsController", () => {
 	beforeEach(() => {
@@ -177,44 +155,6 @@ describe("BlogsController", () => {
 			const response = await request(app).delete("/blogs/" + 1);
 
 			expect(response.status).toBe(204);
-		});
-	});
-
-	describe("GET /blogs/:id/comments", () => {
-		it("should get all blog comments", async () => {
-			(BlogsService.getBlogComments as jest.Mock).mockResolvedValue(mockComments);
-
-			const response = await request(app).get("/blogs/" + 1 + "/comments");
-
-			expect(response.status).toBe(200);
-			expect(response.body).toEqual({
-				data: mockComments.map((mockComment) => ({
-					...mockComment,
-					createdAt: mockComment.createdAt.toISOString(),
-					updatedAt: mockComment.updatedAt.toISOString(),
-				})),
-				message: "Blog comments retrieved successfully",
-			});
-		});
-	});
-
-	describe("POST /blogs/:id/comments", () => {
-		it("should create a blog comment", async () => {
-			(BlogsService.createBlogComment as jest.Mock).mockResolvedValue(mockComment);
-
-			const response = await request(app)
-				.post("/blogs/" + 1 + "/comments")
-				.send({ content: "Comment 1" });
-
-			expect(response.status).toBe(200);
-			expect(response.body).toEqual({
-				data: {
-					...mockComment,
-					createdAt: mockComment.createdAt.toISOString(),
-					updatedAt: mockComment.updatedAt.toISOString(),
-				},
-				message: "Blog comment created successfully",
-			});
 		});
 	});
 });
